Move BillOfSale schema and update references
